backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
beanie==1.25.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastapi==0.110.1
h11==0.14.0
httptools==0.6.1
idna==3.6
lazy-model==0.2.0
motor==3.4.0
numpy==1.26.4
pandas==2.2.1
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
pymongo==4.6.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
toml==0.10.2
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
